+++ BitVector.h
@@ -16,11 +16,11 @@
 typedef struct
 {
 	uint8_t *bits;
-	unsigned int units;
+	size_t units;
 } BitVector;
 
 //initialize the BitVector to hold num_bits number of bits
-void INITIALIZE(BitVector *bv, unsigned int num_bits);
+void INITIALIZE(BitVector *bv, size_t num_bits);
 //free the BitVector
 void FREE_BITVECTOR(BitVector *bv);
 //invert every bit in the BitVector bv
@@ -30,12 +30,12 @@ void SET_ALL_BITS(BitVector *bv);
 //zero the BitVector
 void UNSET_ALL_BITS(BitVector *bv);
 //set the bit at bit_num (with zero-indexing) to 'true' (i.e., 1)
-void SET_BIT(BitVector *bv, unsigned int bit_num);
+void SET_BIT(BitVector *bv, size_t bit_num);
 //set the bit at bit_num  to 'false' (i.e., 0)
-void UNSET_BIT(BitVector *bv, unsigned int bit_num);
+void UNSET_BIT(BitVector *bv, size_t bit_num);
 //toggle the bit in the given position - i.e, 1 goes to 0, 0 to 1
-void INVERT_BIT(BitVector *bv, unsigned int bit_num);
+void INVERT_BIT(BitVector *bv, size_t bit_num);
 //get the truth value of the bit at bit_num, either 0 or 1
-unsigned short GET_BIT(BitVector *bv, unsigned int bit_num);
+unsigned short GET_BIT(BitVector *bv, size_t bit_num);
 
 #endif
